Inverted positions and fixed reported history
The positions as well as the reported histories of the covers were inverted/wrong, which is fixed now. I know it seems counterintuitive in the code, but the reporting within Home Assistant's history is now working correctly and the position usage is in line with the Bosch Smart Home App. Google Assistant works correctly too now.

In addition, any position that is not fully open (0 now being open and 100 being closed) is now getting reported as closed. This makes more sense in my opinion, as shutters are mostly fully open when considered open and often only closed halfway or so, while already being considered (partly) closed.

I was unsure about the underlying architecture's usage of float and int, so I tried to adapt to the style and used what seemed safest. I hope this is correct, if not feel free to change it.

This was briefly tested and worked as expected without any problems at all.